fix: 점수 표시 소수점 제거

+++ client/presentation/src/main/java/com/everyone/movemove_android/ui/watching_video/WatchingVideoScreen.kt
@@ -426,7 +426,7 @@ fun MoveMoveScoreboard(
                     coroutineScope.launch {
                         snackBarState.showSnackbar(
                             message = context.getString(R.string.rating_video_snackbar_message)
-                                .format(rating),
+                                .format(rating.toInt()),
                             duration = SnackbarDuration.Short
                         )
                     }
